refactor db filesistem
